global options, login screen work
This changeset adds an object in the global main that can be used for a single place to enable/disable/modify behavior with one-line changes.

the first optional path now is the starting screen, if login is enabled, it becomes the landing screen, else the intro page with instructions becomes the landing page.

the login page was tweaked for layout/css cleanup, and added some robustness feedback when there is no server to handle GET and POST requests that fail with a locally-hosted development server (no backend scripts or database)