implemented control of AI feedback advice
chevron and ghost car rendering are now independently turned on/off via a application settings.

Bodies in chipmunk can now be tagged as hidden or not; the rendering code checks for this and skips rendering the shapes that belong to that body.

tagging of the bodies with this property is done by the player objects and is controlled by the application options.

more JSDoc comments.